Add concrete classes and enums for options.
Adds the "common interfaces" `IOfN`, `BatchSize`, and `BatchFormat`.
These classes are the exact values that implementations will use,
i.e. an implementation of the SOMA API will not implement its own
`IOfN` or any of these other “option” types.

These types are implemented using the `attrs` package, essentially
an enhanced version of the built-in `dataclasses` package. It is
very lightweight and already extensively used within the Python
ecosystem so it will not add much (if anything) to the size of a typical
installed environment.