Implement `Measurement`, the first of the composed types.
The SOMA composed types are implemented as concrete, instantiable
classes that can be instantiated as wrappers around the
implementation-specific Collection classes.

When loading a composed SOMA object from storage, the "smart open"
function of the implementation will identify the collection as a
composed type, load it into its Collection implementation, and then
wrap it with the appropriate behavior decorator.

    # In an open function:
    collection = load_collection_from_uri(uri)
    # After determining that the collection is an Experiment:
    exp = composed.Experiment(collection)
    # return `exp` to the user.