Harmonize privacy markings on modules and functions.
Certain internal-use-only modules export members which are intended
to be used by other code within somacore. Instead of marking those
exported members as private, this change moves the private marking
to the module itself:

    internal_mod._exported_member
    _internal_mod.exported_member

The member itself is still intended only for internal use, but tooling
can recognize that even if the function (constant, class, whatever)
is unused within the module itself, it is intended for consumption
within somacore.

Also marks `query.AxisIndexer` as public, since it is returned from
a public method (though is not expected to be constructed by user code).