Move to using relative imports wherever possible.
Previously this code had largely been following the Google style guide
recommendation to always use absolute imports (even for things in the
same package). However, that recommendation was developed in the google3
monorepo where everything was effectively one distribution unit and
hierarchies could be extremely deep. In the context of a small,
self-contained package, that recommendation does not make sense.

This change rewrites all local imports to be relative and uses some
strategic re-exporting to simplify the heirarchy.