Declare `soma_type` as a class var rather than a property. (#102)
When using `@property` declarations, the type system would not realize
that in the implementation, `soma_type` could always be accessed as a
class variable. This change just makes `soma_type` a class var and
shuffles around some annotations to make that easier. The `is_sparse`
field on `NDArray` is similarly updated.