Upgrade to mypy v1.0.0 and simplify pre-commit setup. (#117)
This updates pre-commit and the recommended development environment
to use mypy v1.0.0 and regenerates all the `requirements.txt` files
based on this. Also has `ruff` play the role of `isort` since that was
causing issues in a new pre-commit setup (and is no longer needed).

Since we have mypy v1.0.0, removes `_Self` hacks!