Add new supported types and improve precision of coordinates.
This removes `None` from the `SparseNDCoords` type (since we now accept
`()` as the "return everything" value) and separates a singular
`SparseNDCoord` type out from it.

Also introduces a generic `Slice` protocol, since there is no
stdlib-based type annotation for "a slice of a given type". This is,
unfortunately, still not useful at runtime, since `slice` is always
going to resolve to `Slice[Any]`, but the annotation is useful for
specification purposes.